% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear.tools.R
\name{check_single_numeric}
\alias{check_single_numeric}
\title{check whether an object is a single numeric number}
\usage{
check_single_numeric(x, sign = 1)
}
\arguments{
\item{x}{the object to be checked}

\item{sign}{the correct sign: 1 or -1}
}
\value{
a boolean, whehter x is a valid numeric.
}
\description{
check whether an object is a single numeric number
}
\examples{

check_single_numeric(x = nrow(ggplot2::diamonds))

}
\keyword{internal}

