% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear.tools.R
\name{Enter_to_Continue}
\alias{Enter_to_Continue}
\title{Enter_to_Continue: wait your response to continue}
\usage{
Enter_to_Continue(df_input_output = NULL)
}
\arguments{
\item{df_input_output}{data.frame.
df_input_output shall be either NULL or a two column data.frame with characters as values, with first column as what you want to type, and second column as what you want to return.
If it is NULL, then it will return ' Press [enter] to continue; Type [s] to stop'.
See the sample code for the df case.}
}
\value{
Type through keyboard to continue in console.
}
\description{
wait your response to continue
}
\examples{

Enter_to_Continue(rbind(c('small','small data'),c('n','normal'),c('w','weird curve')))

}

