% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear.tools.R
\name{expand_grid}
\alias{expand_grid}
\title{a wrap up of \code{expand.grid()} that takes list}
\usage{
expand_grid(List, KEEP.OUT.ATTRS = TRUE, stringsAsFactors = TRUE)
}
\arguments{
\item{List}{a list same as '...' in \code{expand_grid()},}

\item{KEEP.OUT.ATTRS}{stringsAsFactors, stringsAsFactors see \code{expand_grid()}}
}
\value{
Tsee \code{expand_grid()}
}
\description{
a wrap up of \code{expand.grid()} that takes list
}
\examples{
expand_grid(list(A= c(1:3),b= letters[1:4]))
expand_grid(List = list(c(1:3)))
}
\keyword{internal}

