% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear.tools.R
\name{get_data_from_lm}
\alias{get_data_from_lm}
\title{get raw data from lm or glm}
\usage{
get_data_from_lm(model)
}
\arguments{
\item{modle}{a lm or glm.}
}
\value{
a data.frame used as raw data for the model.
}
\description{
get raw data from lm or glm
}
\examples{

data_used = ggplot2::diamonds[0:10,]
model = lm(price~ cut + carat + I(carat^2) + I(carat^3) +
             I(carat  * depth) + cut:depth, data_used) # a GLM
get_data_from_lm(model)

data_used = 'data_used is deleted in this environment'
get_data_from_lm(model)

}
\keyword{internal}

