% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear.tools.R
\name{get_model_with_coeff}
\alias{get_model_with_coeff}
\title{get a list of model variables with their corresponding coeff vars.}
\usage{
get_model_with_coeff(model, data = NULL)
}
\arguments{
\item{model}{See \code{\link{get_model_pair}}}

\item{data}{See \code{\link{get_model_pair}}}
}
\value{
a list with names as model vars and elements as their corresponding coeff
}
\description{
a wrap up function of \code{\link{get_model_pair}}
}
\details{
See \code{\link{get_model_pair}}
}
\examples{

get_model_with_coeff(price~  I(carat^   2) + cut  + carat*table, data= ggplot2::diamonds)

}

