% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear.tools.R
\name{get_model_with_raw}
\alias{get_model_with_raw}
\title{get a list of model vars with their corresponding raw vars.}
\usage{
get_model_with_raw(model, data = NULL)
}
\arguments{
\item{model, }{See \code{\link{get_model_pair}}}

\item{data, }{See \code{\link{get_model_pair}}}
}
\value{
a list with names as model vars and elements as their raw coeff
}
\description{
a warp up function of \code{\link{get_model_pair}}
}
\details{
See \code{\link{get_model_pair}}
}
\examples{

get_model_with_raw(price~  I(carat^   2) + cut  + carat*table, data= ggplot2::diamonds)

}

