% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear.tools.R
\name{is_decrease}
\alias{is_decrease}
\title{check whether the vector is decreasing}
\usage{
is_decrease(x)
}
\arguments{
\item{x}{a numeric vector}
}
\value{
\code{TRUE} if decreasing, \code{FALSE} if npt.
}
\description{
check whether the vector is decreasing
}
\examples{
is_decrease(rev(c(1,1,1,10,10)))
is_decrease(c(1,2,1,10,10))
}
\keyword{internal}

