% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear.tools.R
\name{is_increase}
\alias{is_increase}
\title{check whether the vector is increasing}
\usage{
is_increase(x)
}
\arguments{
\item{x}{a numeric vector}
}
\value{
\code{TRUE} if increasing, \code{FALSE} if not
}
\description{
check whether the vector is increasing
}
\examples{
is_increase(c(1,1,1,10,10))
is_increase(c(1,2,1,10,10))
}
\keyword{internal}

