% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear.tools.R
\name{Mode}
\alias{Mode}
\title{get mode number from a numeric vector}
\usage{
Mode(x)
}
\arguments{
\item{x}{a numeric vector}
}
\value{
a numeric value mode number of x
}
\description{
get mode number from a numeric vector
}
\examples{
Mode(c(1,1,1,10,10))
}
\keyword{internal}

