% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear.tools.R
\name{stripGlmLR}
\alias{stripGlmLR}
\title{make the lm or glm thin}
\usage{
stripGlmLR(model)
}
\arguments{
\item{model}{glm or lm.}
}
\value{
a thinner model
}
\description{
make the lm or glm thin
}
\author{
Nina Zumel
}
\keyword{internal}

